S_FOLDER/s-w-16/M-N_practice/1013/requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
colorama==0.4.4
decorator==5.1.1
Django==3.2.13
django-appconf==1.0.5
django-extensions==3.2.1
django-imagekit==4.1.0
django-seed==0.3.1
djangorestframework==3.14.0
executing==0.8.3
Faker==15.1.1
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
pickleshare==0.7.5
pilkit==2.0
Pillow==9.2.0
prompt-toolkit==3.0.29
pure-eval==0.2.2
Pygments==2.12.0
python-dateutil==2.8.2
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
stack-data==0.2.0
toposort==1.7
traitlets==5.2.2.post1
wcwidth==0.2.5
